% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcasting_v2.R
\name{nowcasting_v2}
\alias{nowcasting_v2}
\title{Run Nowcasting v2}
\usage{
nowcasting_v2(
  data,
  onset_var,
  report_var,
  time_struc = "rw1",
  family = "zeroinflatednbinomial0",
  allow_time_varying = T,
  delay_struc = "ar",
  delay_ar_order = 2,
  days_forecast = 0,
  max_delay = 15
)
}
\arguments{
\item{data}{The notification system registers data frame. Should be composed in a way that each line represents a notification, with onset and report dates.}

\item{onset_var}{The variable name in \code{data} that corresponds to symptom onset/occurrence date.}

\item{report_var}{The variable name in \code{data} that corresponds to the date when information is registered in the system.}

\item{time_struc}{The temporal random effect structure for \code{inla()}. Should be 'rw1', 'rw2', 'ar1', 'ar2'.}

\item{family}{The model likelihood passed to \code{inla()}}

\item{allow_time_varying}{boolean. Whether to use time-varying delay effects (varying by epidemiological week).}

\item{delay_struc}{Model structure for the delay effect, traditionally "ar".}

\item{delay_ar_order}{In case \code{delay_struc = "ar"}, which order?}

\item{days_forecast}{Number of days in the future to forecast. 0 for only nowcasting.}

\item{max_delay}{Maximum days of delay to account to. Should be based on the distribution of the delays. Raising this parameter increases expressively the computational time and memory used by the function.}
}
\value{
A list of tibbles containing:
\describe{
\item{daily_estimates}{A tibble containing the number of notifications observed and estimated (a.k.a., yet to enter) for each onset date.}
\item{delay_matrix_df}{A tibble specifying the values of the possibly time-varying delay effects.}
\item{week_day_df}{A tibble with the day-of-week effects estimated.}
\item{delay_df}{More detailed tibble of estimated and observed number of notifications by day, with credible 95\% intervals.}
}
}
\description{
Run nowcasting function for notification system based data with informed parameters.
}
\examples{
# Using built-in data from Dengue notifications in Rio
ncast_results <- dengue_rio |> nowcasting_v2(onset_var = "DT_SIN_PRI", report_var = "DT_DIGITA")

if (requireNamespace("ggplot2", quietly = TRUE) &&
    requireNamespace("magrittr", quietly = TRUE)) {
ncast_results$daily_estimates |>
 ggplot2::ggplot(ggplot2::aes(x=onset, y=n)) +
 ggplot2::geom_bar(ggplot2::aes(fill = observed), stat = "identity", position = "stack")
}

}
